requirement.txt
mysql-connector-python
transformers
torch
ctransformers


accelerate
sentencepiece
protobuf
